+++ src/components/sdk/MeetingControls/AudioInputVFControl.tsx
@@ -49,7 +49,7 @@ const AudioInputVFControl: React.FC<Props> = ({
   const meetingManager = useMeetingManager();
   const [isLoading, setIsLoading] = useState(false);
   // When the user click on Amazon Voice Focus option, the state will change.
-  const [isVoiceFocusOn, setIsVoiceFocusOn] = useState(false);
+  const [isVoiceFocusChecked, setIsVoiceFocusChecked] = useState(false);
   // Only when the current input audio device is an Amazon Voice Focus device, the state will be true. Otherwise, it will be false.
   const [isVoiceFocusEnabled, setIsVoiceFocusEnabled] = useState(false);
   const [dropdownWithVFOptions, setDropdownWithVFOptions] = useState<
@@ -75,6 +75,10 @@ const AudioInputVFControl: React.FC<Props> = ({
   }, []);
 
   useEffect(() => {
+    console.info(`Amazon Voice Focus is ${isVoiceFocusEnabled ? 'enabled' : 'disabled'}.`);
+  }, [isVoiceFocusEnabled]);
+
+  useEffect(() => {
     // Only when the current input audio device is an Amazon Voice Focus transform device,
     // Amazon Voice Focus button will be checked.
     if (device instanceof VoiceFocusTransformDevice) {
@@ -91,7 +95,7 @@ const AudioInputVFControl: React.FC<Props> = ({
         children={<span>{device.label}</span>}
         checked={isOptionActive(selectedDevice, device.deviceId)}
         onClick={async (): Promise<void> => {
-          if (isVoiceFocusOn && !isLoading) {
+          if (isVoiceFocusChecked && !isLoading) {
             setIsLoading(true);
             const receivedDevice = audioInputSelectionToDevice(device.deviceId);
             const currentDevice = await addVoiceFocus(receivedDevice);
@@ -118,11 +122,10 @@ const AudioInputVFControl: React.FC<Props> = ({
           disabled={isLoading}
           onClick={() => {
             setIsLoading(true);
-            setIsVoiceFocusOn((current) => !current);
-          }}
-        />
-      );
-      dropdownOptions?.push(<PopOverSeparator key="separator" />);
+            setIsVoiceFocusChecked(current => !current);
+          }} />
+      )
+      dropdownOptions?.push(<PopOverSeparator key = 'separator' />);
       dropdownOptions?.push(vfOption);
     }
 
@@ -137,7 +140,7 @@ const AudioInputVFControl: React.FC<Props> = ({
     devices[0].label,
     isLoading,
     isVoiceFocusEnabled,
-    isVoiceFocusOn,
+    isVoiceFocusChecked,
     isVoiceFocusSupported,
     selectedDevice,
   ]);
@@ -145,12 +148,14 @@ const AudioInputVFControl: React.FC<Props> = ({
   useEffect(() => {
     async function onVFCheckboxChange() {
       let current = device;
-      if (isVoiceFocusOn) {
-        if (typeof device === 'string') {
+      if (isVoiceFocusChecked) {
+        console.info('User turned on Amazon Voice Focus.')
+        if (typeof (device) === 'string') {
           const currentDevice = audioInputSelectionToDevice(device);
           current = await addVoiceFocus(currentDevice);
         }
       } else {
+        console.info('Amazon Voice Focus is off by default or user turned off Amazon Voice Focus.')
         if (device instanceof VoiceFocusTransformDevice) {
           current = device.getInnerDevice();
         }
@@ -160,7 +165,7 @@ const AudioInputVFControl: React.FC<Props> = ({
     }
 
     onVFCheckboxChange();
-  }, [isVoiceFocusOn]);
+  }, [isVoiceFocusChecked]);
 
   return (
     <ControlBarButton