+++ src/providers/VoiceFocusProvider/index.tsx
@@ -71,11 +71,12 @@ const VoiceFocusProvider: React.FC<Props> = ({ spec, options, children }) => {
     }
   );
 
-  const addVoiceFocus = async (
-    device: Device
-  ): Promise<Device | VoiceFocusTransformDevice> => {
+  const addVoiceFocus = async (device: Device): Promise<Device | VoiceFocusTransformDevice> => {
+    console.info('Add Amazon Voice Focus to the following audio input device', device);
+
     if (voiceFocusDevice) {
       const vf = await voiceFocusDevice.chooseNewInnerDevice(device);
+      console.info('Re-used the same internal state to create an Amazon Voice Focus transform device.');
       setVoiceFocusDevice(vf);
       return vf;
     }
@@ -89,6 +90,7 @@ const VoiceFocusProvider: React.FC<Props> = ({ spec, options, children }) => {
       const transformer = await getVoiceFocusDeviceTransformer();
       const vf = await transformer?.createTransformDevice(device);
       if (vf) {
+        console.info('Created a new Amazon Voice Focus transform device.');
         setVoiceFocusDevice(vf);
         return vf;
       }
@@ -179,11 +181,17 @@ const VoiceFocusProvider: React.FC<Props> = ({ spec, options, children }) => {
     }
 
     if (isVoiceFocusSupported) {
-      console.debug('Amazon Voice Focus is supported.');
+      console.info('Amazon Voice Focus is supported.');
     } else {
       console.warn('Amazon Voice Focus is not supported.');
     }
   }, [isVoiceFocusSupported]);
+
+  useEffect(() => {
+    if (voiceFocusDevice) {
+      console.info('Current Amazon Voice Focus transform device: ', voiceFocusDevice);
+    } 
+  }, [voiceFocusDevice]);
 
   const value: VoiceFocusState = {
     isVoiceFocusSupported,